common/vault/requirements.txt
hvac
python-dotenv
watchdog